+++ triplea/cli/import.py
@@ -1,0 +1,21 @@
+import click
+from triplea.service.click_logger import logger
+from triplea.service.repository.state.initial import get_article_from_bibliography_file_format
+
+@click.command()
+@click.argument('filename',
+                type=click.Path(exists=True)
+                 )
+def import_single_file(filename):
+    
+    logger.INFO(f'Import file type : .bib , .enw , .ris')
+    r = get_article_from_bibliography_file_format(filename)
+    if r:
+        logger.INFO(f'Import complete.')
+    else:
+        logger.ERROR(f'Import incomplete.')
+
+
+
+if __name__ == '__main__':
+    import_single_file()