+++ triplea/cli/ner.py
@@ -1,6 +1,6 @@
 from string import printable
 import click
-from triplea.service.nlp.ner_training import get_title_ner
+from triplea.service.nlp.ner import get_title_ner
 from triplea.service.click_logger import logger
 
 @click.command()